Add new features to MATLAB interface
The C interface now takes a recursion method for the Wigner functions
just like SSHT does (and the value simply gets passed through to SSHT).
Additionally, the MATLAB interface has been extended to support this new
parameter as well as a flag for real signals and the n-mode optimisation
options. A new demo has been added for the case of only even n, and the
demo for real signals now uses the appropriate methods.