Add DFT option to inverse transforms
Inverse transforms can now optionally perform the Fourier transform over
n/gamma as a manual DFT instead of using FFTW. This is not reflected in
the Matlab interface yet.

So far no cases have been found where the DFT is faster than FFTW, which
is why this is moved to a branch, which won't be pursued further for
now.